fix constructor param type

+++ src/animal/Fox.cpp
@@ -12,7 +12,7 @@ Fox::Fox(uint16_t x, uint16_t y, World *world) : Animal(x, y, world) {
     initiative = 7;
 }
 
-Fox::Fox(uint16_t x, uint16_t y, uint8_t power, uint8_t initiative, World *world)
+Fox::Fox(uint16_t x, uint16_t y, uint8_t power, uint16_t initiative, World *world)
         : Animal(x, y, power, initiative, world) {
     sign = 'F';
     this->initiative = 7;
